limit int length

+++ src/db_models/steps.py
@@ -6,7 +6,7 @@ from sqlalchemy.types import Integer
 class StepsTable(SQLModel, table=True):
     __tablename__ = "steps"
 
-    id: Optional[int] = Field(default=None, sa_column=Column(Integer, primary_key=True))
+    id: Optional[int] = Field(default=None, sa_column=Column(Integer(6), primary_key=True))
     position: int
     image_url: Optional[str] = Field(default=None)
     description: str
